Moved DbConnectionFactory to the Infrastructure project
I moved the implementation of IDbConnection to its own file and to the infrastructure project to comply with the clean architecure.
