+++ Cookbook.Infrastructur/InfrastructureServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Cookbook.Application;
+using Cookbook.Application.Database;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Cookbook.Infrastructur
@@ -8,6 +9,13 @@ namespace Cookbook.Infrastructur
         public static IServiceCollection AddPersistence(this IServiceCollection services)
         {
             services.AddSingleton<ICookbookRepository, CookbookRepository>();
+
+            return services;
+        }
+
+        public static IServiceCollection AddDatabase(this IServiceCollection services)
+        {
+            services.AddSingleton<IDbConnectionFactory>(_ => new SqliteConnectionFactory());
 
             return services;
         }
